+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/CraftingPreview.java
@@ -1,6 +1,0 @@
-package com.refinedmods.refinedstorage.api.autocrafting;
-
-import com.refinedmods.refinedstorage.api.resource.list.ResourceList;
-
-public record CraftingPreview(ResourceList missing, ResourceList toTake, ResourceList toCraft) {
-}