+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/screen/GridSearchBoxWidget.java
@@ -12,7 +12,6 @@ import com.refinedmods.refinedstorage.query.lexer.SyntaxHighlighter;
 
 import java.util.HashSet;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 import java.util.function.Consumer;